file: refine filetype detection
The eight file-types based on CFB or ZIP compound files can be detected
with 100% accuracy (or at least 99.999%). Try this strategy first,
ignoring the unreliable content-type for these file-types.

This fixes a problem where a CSV file with an asserted XLS file-type was
mistakenly typed as XLS and failed partitioning.