fix(ci): Use non-deprecated way of invoking ruff in `make tidy` (#3825)
I noticed that `make tidy` wasn't working in my development environment.
This happens if you, a developer, forget to follow the specific
instructions in `README.md` and install exactly the right versions of
the necessary tools, including a *quite old* version of Ruff. This
version will nonetheless warn you:

warning: `ruff <path>` is deprecated. Use `ruff check <path>` instead.

So this fixes that, in order to future-proof and avoid confusion!