fix/Prepare for breaking change in upcoming unstructured-client
In the unreleased 0.30.0 of unstructured-client, handling of the server_url is changing. Instead of
passing a url to the instance of the client, it's preferred that we set the server_url per
endpoint. We can make this change against existing versions of the client in order to be compatible
when 0.30.0 is released.