email date format flexibility (#4072)
we are seeing some .eml files come through the VLM partitioner. Which
then downgrades to hi-res i believe.

For some reason they have a date format that is not standard email
format. But it is still legitimate.

This uses a more robust date package to parse the date. This package is
already installed.

---------

Co-authored-by: ryannikolaidis <1208590+ryannikolaidis@users.noreply.github.com>
Co-authored-by: potter-potter <potter-potter@users.noreply.github.com>