Setup Codeflash Github Actions to optimize all future code (#4082)
- This Pull Request sets up the `codeflash.yml` file which will run on
every new Pull Request that modifies the source code for `unstructured`
directory.
- We setup the codeflash config in the pyproject.toml file. This defines
basic project config for codeflash.
- The workflow uses uv to install the CI dependencies faster than your
current caching solution. Speed is useful to get quicker optimizations.
- Please take a look at the requirements that are being installed. Feel
free to add more to the install list. Codeflash tries to execute code
and if it is missing a dependency needed to make something run, it will
fail to optimize.
- Codeflash is being installed everytime in the CI. This helps the
workflow always use the latest version of codeflash as it improves
rapidly. Feel free to add codeflash to dev dependency as well, since we
are about to release more local optimization tools like VS Code and
claude code extensions.
- Feel free to modify this Github action anyway you want

**Actions Required to make this work-**

- Install the Codeflash Github app from [this
link](https://github.com/apps/codeflash-ai/installations/select_target)
to this repo. This is required for our github-bot to comment and create
suggestions on the github repo.
- Create a new `CODEFLASH_API_KEY` after signing up to [Codeflash from
our website](https://www.codeflash.ai/). The onboarding will ask you to
create an API Key and show instructions on how to save the api key on
your repo secrets.

Then, after this PR is merged in it will start generating new
optimizations 🎉

---------

Signed-off-by: Saurabh Misra <misra.saurabh1@gmail.com>
Co-authored-by: Aseem Saxena <aseem.bits@gmail.com>
Co-authored-by: cragwolfe <cragcw@gmail.com>